Safari-611.1.10/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-invalid.html

<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
        testRunner.dumpResourceLoadCallbacks();
    }
</script>

<p>The frame below should load, and a console message should be generated
that notes the invalid header.</p>
<iframe style="width:500px; height:500px" src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi"></iframe>
