LayoutTests/platform/chromium/http/tests/security/XFrameOptions/x-frame-options-invalid-expected.txt

http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi, main document URL http://127.0.0.1:8000/security/XFrameOptions/x-frame-options-invalid.html, http method GET> redirectResponse (null)
CONSOLE MESSAGE: Invalid 'X-Frame-Options' header encountered when loading 'http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi': 'INVALID INVALID INVALID' is not a recognized directive. The header will be ignored.
http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi - didReceiveResponse <NSURLResponse http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi, http status code 200>
http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-invalid.cgi - didFinishLoading
The frame below should load, and a console message should be generated that notes the invalid header.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: This text should show up.
